Chisel Service Health Check
This adds a simple endpoint (nothing more than a 200 OK) on /health to check the chisel service is running
